refactor: split scanner and add rename detection
