+++ src/sample_sources/scanner/mod.rs
@@ -1,5 +1,6 @@
-mod metadata;
 mod scan;
+mod scan_diff;
+mod scan_fs;
 
 pub use scan::{
     scan_in_background, scan_once, scan_with_progress, hard_rescan, ChangedSample, ScanError,
